[node] Split `build()`, `prepareCache()` and `startDevServer()` into separate files
Small tidiness refactor. The old `src/index.ts` was getting a bit unweildy.
