+++ .changeset/tough-terms-burn.md
@@ -1,0 +1,5 @@
+---
+'@vercel/node': patch
+---
+
+Split `build()`, `prepareCache()` and `startDevServer()` into separate files
