+++ packages/node/src/prepare-cache.ts
@@ -1,0 +1,5 @@
+import { glob, type PrepareCache } from '@vercel/build-utils';
+
+export const prepareCache: PrepareCache = ({ repoRootPath, workPath }) => {
+  return glob('**/node_modules/**', repoRootPath || workPath);
+};
